Specify client and server ports directly, not full go style addresses
Instead of listening on say ":15000" we not just tell them to listen on
port "15000".  This makes the command arguments simpler, and we can control
where connections are accepted from in a separate way.

+++ core.go
@@ -8,7 +8,7 @@ import (
 
 var (
 	// Port used for network communications
-	DefaultPort = 15796
+	DefaultPort = uint(15796)
 	// Whether or not we have debug logging on
 	DebugLogging = false
 )
